% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.FRBS.R
\name{Create.FRBS.R}
\alias{Create.FRBS.R}
\title{Function to create a zero-order Takagi–Sugeno–Kang fuzzy rule-based systems (FRBS)}
\usage{
Create.FRBS.R(ImpVariables, Range, MFfunction, MFparameters, Consequents)
}
\arguments{
\item{ImpVariables}{the names of the input variables in vector format}

\item{Range}{a matrix with as many columns as input variables with the minimum and maximum expected values for each variable}

\item{MFfunction}{the selected membership function}

\item{MFparameters}{the parameters of the membership functions. It should be a matrix with as many columns as fuzzy sets. The columns for each variable must include the variable name as indicated in \emph{ImpVariables}.}

\item{Consequents}{the zero-order TSK fuzzy rule consequents.}
}
\value{
a list containing the five components of the FRBSs
}
\description{
Function to create a zero-order Takagi–Sugeno–Kang fuzzy rule-based systems (FRBS)
}
\examples{
FRBS <- Create.FRBS.R(ImpVariables = c("Velocity",
                                       "Depth",
                                       "Substrate.index",
                                       "Cover.index"),
                      Range = matrix(c(0, 1.164, 0, 2.67, 0, 8, 0, 1), nrow = 2, byrow = FALSE,
                                     dimnames = list(NULL, c("Velocity",
                                                             "Depth",
                                                             "Substrate.index",
                                                             "Cover.index"))),
                      MFfunction = PIMF,
                      MFparameters = matrix(c(0, 0, 0.0377, 0.4,
                                              0.0377, 0.4, 1.164, 1.164,
                                              0, 0, 0.46, 0.8767,
                                              0.46, 0.8767, 1.3067, 2.67,
                                              1.3067, 2.67, 2.67, 2.67,
                                              0, 0, 0, 0.5333,
                                              0, 0.5333, 3.8667, 6,
                                              3.8667, 6, 8, 8,
                                              0, 0, 0, 1,
                                              0, 1, 1, 1), nrow = 4, byrow = FALSE,
                                     dimnames = list(NULL, c("VelocityL", "VelocityH",
                                            "DepthL", "DepthM", "DepthH",
                                            "Substrate.indexL", "Substrate.indexM",
                                            "Substrate.indexH",
                                            "Cover.indexL", "Cover.indexH"))),
                      Consequents = c(0, 0, 1, 0,
                      1, 0, 0, 0,
                      1, 0, 0, 0,
                      0, 0, 0, 0,
                      0, 0, 1, 0,
                      1, 1, 0, 0.5,
                      0, 0, 1, 0,
                      0, 1, 1, 0,
                      0, 0, 0, 0)
                      )

}
