% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateFIS_cpp.R
\name{evaluateFIS_cpp}
\alias{evaluateFIS_cpp}
\title{Evaluate Fuzzy Inference System (FIS)}
\usage{
evaluateFIS_cpp(
  inputData,
  membershipFunctions,
  ruleConsequents,
  inputLimits = NULL
)
}
\arguments{
\item{inputData}{A data frame or matrix containing the input data to be evaluated by the FIS.}

\item{membershipFunctions}{A list or matrix of fuzzy membership functions corresponding to the input data.}

\item{ruleConsequents}{A list or matrix of fuzzy rule consequents, which are the results of the rules applied to the input data.}

\item{inputLimits}{(optional) A list or vector specifying the limits for the input data. Default is \code{NULL}.}
}
\value{
A numeric vector of output values resulting from the evaluation of the fuzzy inference system.
}
\description{
This function evaluates a fuzzy inference system (FIS) based on the provided input data,
membership functions, and rule consequents. It processes the inputs using fuzzy logic and
returns the evaluated output.
}
\examples{
# Example usage:
result <- evaluateFIS_cpp(inputData, membershipFunctions, ruleConsequents, inputLimits)
}
