% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREDICT.FRBS.Fast.R
\name{FUZZIFY.FRBS.Fast}
\alias{FUZZIFY.FRBS.Fast}
\title{Function to obtain the membership of each datum to each fuzzy rule}
\usage{
FUZZIFY.FRBS.Fast(Data, FRBS)
}
\arguments{
\item{Data}{the the dataset in data.frame format.}

\item{FRBS}{the list compiling the elements describing a zero-order Takagi–Sugeno–Kang fuzzy rule-based system.}
}
\value{
matrix with the membership of each datum to each fuzzy rule. Rows are fuzzy rules and columns each datum in the dataset.
}
\description{
Function to obtain the membership of each datum to each fuzzy rule
}
\examples{
Memberships <- FUZZIFY.FRBS.Fast(Data =
Lepomis.gibbosus.dataset[,Lepomis.gibbosus.FRBS$ImpVariables],
FRBS = Lepomis.gibbosus.FRBS)
print(Memberships[, 1:12])

}
