% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyFishHSRCpp-package.R
\docType{package}
\name{FuzzyFishHSRCpp-package}
\alias{FuzzyFishHSRCpp}
\alias{FuzzyFishHSRCpp-package}
\title{Package-level documentation}
\description{
FuzzyFishHSRCpp provides functions to develop and inspect zero-order presence-absence Takagi–Sugeno–Kang fuzzy rule-based systems for fish habitat evaluation within environmental flow assessemnt studies. The main difference with the former packages relies in the function to carry out predictions, which has been reprogremmed in C++.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RafaMMas/FuzzyFishHSRCpp}
  \item Report bugs at \url{https://github.com/RafaMMas/FuzzyFishHSRCpp/issues}
}

}
\author{
\strong{Maintainer}: Muñoz-Mas Rafael \email{rafa.m.mas@gmail.com} (\href{https://orcid.org/0000-0002-7263-5799}{ORCID})

}
\keyword{internal}
