% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembershipFunctions.R
\name{GAUSSMF}
\alias{GAUSSMF}
\title{Gaussian membership function}
\usage{
GAUSSMF(pattern, parameters)
}
\arguments{
\item{pattern}{the values of the input variable for whom the membership is calculated}

\item{parameters}{the parameters of the membership function}
}
\value{
membership of the evaluated values
}
\description{
Gaussian membership function
}
\examples{
plot(GAUSSMF(pattern = seq(0, 100, by = 1), parameters = c(35, 5)),
type = "l", xlab = "Value", ylab = "Membership", bty = "n", las = 1, col = "orangered")

plot(GAUSSMF(pattern = seq(0, 100, by = 1), parameters = c(60, 25)),
type = "l", xlab = "Value", ylab = "Membership", bty = "n", las = 1, col = "orangered")

}
