% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembershipFunctions.R
\name{GBELLMF}
\alias{GBELLMF}
\title{Gaussian bell membership function}
\usage{
GBELLMF(pattern, parameters)
}
\arguments{
\item{pattern}{the values of the input variable for whom the membership is calculated}

\item{parameters}{the parameters of the membership function}
}
\value{
membership of the evaluated values
}
\description{
Gaussian bell membership function
}
\examples{
plot(GBELLMF(pattern = seq(0, 100, by = 1), parameters = c(10, 1, 50)),
type = "l", xlab = "Value", ylab = "Membership", bty = "n", las = 1, col = "orangered")

plot(GBELLMF(pattern = seq(0, 100, by = 1), parameters = c(20, 5, 50)),
type = "l", xlab = "Value", ylab = "Membership", bty = "n", las = 1, col = "orangered")

}
