% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{list}
\name{Lepomis.gibbosus.FRBS}
\alias{Lepomis.gibbosus.FRBS}
\title{Optimal Fuzzy Rule-Based System (FRBS) Model for \emph{Lepomis gibbosus}

This object represents the optimal FRBS model developed for predicting the habitat suitability
of \emph{Lepomis gibbosus} based velocity, depth, substrate and cover.}
\usage{
data(Lepomis.gibbosus.FRBS)
}
\description{
Optimal Fuzzy Rule-Based System (FRBS) Model for \emph{Lepomis gibbosus}

This object represents the optimal FRBS model developed for predicting the habitat suitability
of \emph{Lepomis gibbosus} based velocity, depth, substrate and cover.
}
\references{
Mu\ifelse{html}{\out{&ntilde;}}{\enc{UTF-8}{ñ}}oz-Mas, R., Macian-Sorribes, H., Oliva-Paterna, F. J.,
Sangelantoni, L., Peano, D., Pulido-Velazquez, M., & Mart\ifelse{html}{\out{&iacute;}}{\enc{UTF-8}{í}}nez-Capel, F. (2024).
Adaptation measures to global change in the Serpis River Basin (Spain): An evaluation considering agricultural benefits,
environmental flows, and invasive fishes. \emph{Ecological Indicators}, 161, 111979.
DOI: 10.1016/j.ecolind.2024.111979
}
\keyword{datasets}
