% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMF.R
\name{PlotMF}
\alias{PlotMF}
\title{Function to plot the membership functions}
\usage{
PlotMF(FRBS, n.pt = 9999, data = NULL, Title = NULL)
}
\arguments{
\item{FRBS}{the list compiling the elements describing a zero-order Takagi–Sugeno–Kang fuzzy rule-based system.}

\item{n.pt}{number of points/resolution used to plot each membership function.}

\item{data}{the training datataset in \emph{data.frame} format.}

\item{Title}{the figure title.}
}
\value{
a 4 by 4 plot of the membership functions for each input variable
}
\description{
Function to plot the membership functions
}
\examples{

data(Lepomis.gibbosus.dataset)  
data(Lepomis.gibbosus.FRBS)  

PlotMF(FRBS = Lepomis.gibbosus.FRBS, n.pt = 9999,
data = NULL,
Title = "Lepomis gibbosus")

PlotMF(FRBS = Lepomis.gibbosus.FRBS, n.pt = 9999,
data = Lepomis.gibbosus.dataset,
Title = "Lepomis gibbosus")

}
