% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zero.order.TSK.plots.R
\name{plotVariableMFs}
\alias{plotVariableMFs}
\title{Plot Membership Functions for a Single Variable}
\usage{
plotVariableMFs(
  mfList,
  variableName = "",
  xlim = c(0, 1),
  n = 100,
  colors = NULL,
  lwd = 2,
  main = NULL,
  xlab = "x",
  ylab = "Membership Degree",
  ...
)
}
\arguments{
\item{mfList}{A list of membership functions. Each element should be a list
with components \code{type} (character) and \code{params} (numeric vector).}

\item{variableName}{Character string. Optional name for the variable being plotted.}

\item{xlim}{Numeric vector of length 2. Range of the x-axis (default is \code{c(0, 1)}).}

\item{n}{Integer. Number of evaluation points along the x-axis (default is 100).}

\item{colors}{Optional character vector of colors to use for the functions.}

\item{lwd}{Numeric. Line width for the membership function curves (default is 2).}

\item{main}{Character. Plot title. If \code{NULL}, a title will be generated automatically.}

\item{xlab}{Character. Label for the x-axis.}

\item{ylab}{Character. Label for the y-axis.}

\item{...}{Additional graphical parameters passed to \code{\link[graphics]{plot}}.}
}
\value{
A plot of membership functions for one variable.
}
\description{
Plot all membership functions defined for a single input variable.
}
\details{
This function helps visualize fuzzy membership functions of a variable using
base R plotting. Each curve is labeled in a legend according to its type and parameters.
}
\examples{
mfList <- list(
  list(type = "trimf", params = c(0, 0.5, 1)),
  list(type = "gaussmf", params = c(0.5, 0.1))
)
plotVariableMFs(mfList, variableName = "Example Variable")

}
