% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREDICT.FRBS.Fast.R
\name{PREDICT.FRBS.Fast}
\alias{PREDICT.FRBS.Fast}
\title{Function to carry out predictions with zero-order Takagi–Sugeno–Kang fuzzy rule-based systems for fish habitat evaluation.}
\usage{
PREDICT.FRBS.Fast(Data, FRBS)
}
\arguments{
\item{Data}{the evaluated dataset in data.frame format}

\item{FRBS}{the list compiling the elements describing a zero-order Takagi–Sugeno–Kang fuzzy rule-based system}
}
\value{
matrix with the membership of each datum to each fuzzy rule obtained using the function FUZZIFY.FRBS.Fast and the predicted habitat suitability in the last column
}
\description{
Function to carry out predictions with zero-order Takagi–Sugeno–Kang fuzzy rule-based systems for fish habitat evaluation.
}
\examples{

data(Lepomis.gibbosus.dataset)  
data(Lepomis.gibbosus.FRBS)

Predict <- PREDICT.FRBS.Fast(Data =
Lepomis.gibbosus.dataset[,Lepomis.gibbosus.FRBS$ImpVariables],
FRBS = Lepomis.gibbosus.FRBS)
plot(Lepomis.gibbosus.dataset$Species, Predict[2,],
bty = "n", las = 1, col = "orangered", xlab = "Observed", ylab = "Predicted")
abline(lm(Predict[2,] ~ Lepomis.gibbosus.dataset$Species), col = "dodgerblue")

}
