% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembershipFunctions.R
\name{PSIGMF}
\alias{PSIGMF}
\title{Product of two sigmoidal membership functions}
\usage{
PSIGMF(pattern, parameters)
}
\arguments{
\item{pattern}{the values of the input variable for whom the membership is calculated}

\item{parameters}{the parameters of the membership function}
}
\value{
membership of the evaluated values
}
\description{
Product of two sigmoidal membership functions
}
\examples{
plot(PSIGMF(pattern = seq(0, 100, by = 1), parameters = c(2, 15, -0.5, 60)),
type = "l", xlab = "Value", ylab = "Membership", bty = "n", las = 1, col = "orangered")

plot(PSIGMF(pattern = seq(0, 100, by = 1), parameters = c(1, 50, -0.5, 75)),
type = "l", xlab = "Value", ylab = "Membership", bty = "n", las = 1, col = "orangered")

}
