% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembershipFunctions.R
\name{SIGMF}
\alias{SIGMF}
\title{Sigmoidal membership functions}
\usage{
SIGMF(pattern, parameters)
}
\arguments{
\item{pattern}{the values of the input variable for whom the membership is calculated}

\item{parameters}{the parameters of the membership function}
}
\value{
membership of the evaluated values
}
\description{
Sigmoidal membership functions
}
\examples{
plot(SIGMF(pattern = seq(0, 100, by = 1), parameters = c(0.5,50)),
type = "l", xlab = "Value", ylab = "Membership", bty = "n", las = 1, col = "orangered")

plot(SIGMF(pattern = seq(0, 100, by = 1), parameters = c(-0.2, 35)),
type = "l", xlab = "Value", ylab = "Membership", bty = "n", las = 1, col = "orangered")

}
