% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembershipFunctions.R
\name{SMF}
\alias{SMF}
\title{S-shaped membership function}
\usage{
SMF(pattern, parameters)
}
\arguments{
\item{pattern}{the values of the input variable for whom the membership is calculated}

\item{parameters}{the parameters of the membership function}
}
\value{
membership of the evaluated values
}
\description{
S-shaped membership function
}
\examples{
plot(SMF(pattern = seq(0, 100, by = 1), parameters = c(10, 60)),
type = "l", xlab = "Value", ylab = "Membership", bty = "n", las = 1, col = "orangered")

plot(SMF(pattern = seq(0, 100, by = 1), parameters = c(0.5, 90)),
type = "l", xlab = "Value", ylab = "Membership", bty = "n", las = 1, col = "orangered")

}
