% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembershipFunctions.R
\name{TRAPMF}
\alias{TRAPMF}
\title{Trapezoidal membership function}
\usage{
TRAPMF(pattern, parameters)
}
\arguments{
\item{pattern}{the values of the input variable for whom the membership is calculated}

\item{parameters}{the parameters of the membership function}
}
\value{
membership of the evaluated values
}
\description{
Trapezoidal membership function
}
\examples{
plot(TRAPMF(pattern = seq(0, 100, by = 1), parameters = c(10, 60, 75, 90)),
type = "l", xlab = "Value", ylab = "Membership", bty = "n", las = 1, col = "orangered")

plot(TRAPMF(pattern = seq(0, 100, by = 1), parameters = c(10, 40, 50, 80)),
type = "l", xlab = "Value", ylab = "Membership", bty = "n", las = 1, col = "orangered")

}
