% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MembershipFunctions.R
\name{TRIMF}
\alias{TRIMF}
\title{Triangular membership function}
\usage{
TRIMF(pattern, parameters)
}
\arguments{
\item{pattern}{the values of the input variable for whom the membership is calculated}

\item{parameters}{the parameters of the membership function}
}
\value{
membership of the evaluated values
}
\description{
Triangular membership function
}
\examples{
plot(TRIMF(pattern = seq(0, 100, by = 1), parameters = c(10, 60, 75)),
type = "l", xlab = "Value", ylab = "Membership", bty = "n", las = 1, col = "orangered")

plot(TRIMF(pattern = seq(0, 100, by = 1), parameters = c(10, 40, 70)),
type = "l", xlab = "Value", ylab = "Membership", bty = "n", las = 1, col = "orangered")

}
