% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.FRBS.R
\docType{data}
\name{Lepomis.gibbosus.dataset}
\alias{Lepomis.gibbosus.dataset}
\title{Lepomis gibbosus dataset continaing four microhabitat variables: Velocity, Depth, Substrate and Cover and its presence (1) and absneec (0).}
\format{
An object of class \code{data.frame} with 1293 rows and 11 columns.
}
\usage{
Lepomis.gibbosus.dataset
}
\description{
Lepomis gibbosus dataset continaing four microhabitat variables: Velocity, Depth, Substrate and Cover and its presence (1) and absneec (0).
}
\keyword{datasets}
