% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.FRBS.R
\docType{data}
\name{Lepomis.gibbosus.FRBS}
\alias{Lepomis.gibbosus.FRBS}
\title{Optimal Fuzzy Rule-Based System (FRBS) Model for \emph{Lepomis gibbosus}

This object represents the optimal FRBS model developed for predicting the habitat suitability
of \emph{Lepomis gibbosus} based on key environmental variables.}
\usage{
data(Lepomis.gibbosus.FRBS)
}
\description{
Optimal Fuzzy Rule-Based System (FRBS) Model for \emph{Lepomis gibbosus}

This object represents the optimal FRBS model developed for predicting the habitat suitability
of \emph{Lepomis gibbosus} based on key environmental variables.
}
\keyword{datasets}
