% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.FRBS.R
\docType{data}
\name{Lepomis.gibbosus.FRBS}
\alias{Lepomis.gibbosus.FRBS}
\title{Optimal model Fuzzy Rule Based System (FRBS) for Lepomis gibbosus}
\format{
An object of class \code{list} of length 5.
}
\usage{
Lepomis.gibbosus.FRBS
}
\description{
Optimal model Fuzzy Rule Based System (FRBS) for Lepomis gibbosus
}
\keyword{datasets}
