% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zero.order.TSK.plots.R
\name{plotAllMFs}
\alias{plotAllMFs}
\title{Plot Membership Functions for All Variables}
\usage{
plotAllMFs(
  membershipFunctions,
  varNames = NULL,
  xlim = c(0, 1),
  n = 100,
  colors = NULL,
  lwd = 2,
  main = NULL,
  xlab = "x",
  ylab = "Membership",
  ...
)
}
\arguments{
\item{membershipFunctions}{A list of lists. Each element represents one variable,
and should be a list of membership functions as used in \code{\link{plotVariableMFs}}.}

\item{varNames}{Optional character vector with names for each variable.}

\item{xlim}{Numeric vector of length 2. Common x-axis range for all plots.}

\item{n}{Integer. Number of evaluation points (default is 100).}

\item{colors}{Optional character vector of colors.}

\item{lwd}{Numeric. Line width for plotted curves.}

\item{main}{Character. Prefix for titles in each plot.}

\item{xlab}{Character. X-axis label.}

\item{ylab}{Character. Y-axis label.}

\item{...}{Additional graphical parameters passed to \code{\link{plotVariableMFs}}.}
}
\value{
A grid of plots, one for each variable.
}
\description{
Plot membership functions for multiple variables, each in a separate panel.
}
\details{
This function visualizes fuzzy membership functions grouped by variable.
It arranges the plots in a grid layout using base R graphics.
}
\examples{
mf1 <- list(
  list(type = "trimf", params = c(0, 0.5, 1)),
  list(type = "gaussmf", params = c(0.5, 0.1))
)
mf2 <- list(
  list(type = "trapmf", params = c(0, 0.3, 0.7, 1)),
  list(type = "sigmf", params = c(10, 0.5))
)
plotAllMFs(list(mf1, mf2), varNames = c("X1", "X2"))

}
